Platform/96Boards/Secure96Dxe: remove /plugin/ from .dts
Even though the Secure96 device tree source is strictly an overlay, we
managed to express it in a way that does not rely on unresolved symbols
and other tricks that are only implemented in fairly recent versions of
the device tree compiler, and so adding the /plugin/ directive is not
only unnecessary, it is harmful because it is only understood by those
same recent compiler versions. So remove it.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>