Silicon/SynQuacerPlatformFlashAccessLib: fix return value on no FVB found
If no suitable FVB protocol implementation is found to apply the capsule
update, and the last one we disregarded was ruled out because it has the
read-only attribute, we will exit the function returning EFI_SUCCESS
without assigning a value to *OutFvb, resulting in a crash when it
subsequently gets dereferenced. So set the correct value for Status for
that case.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>