Platform/ARM: Fix NOR Flash dependency for CTA15A7
The correct load order for the NOR Flash driver and Runtime Variables
to work is as below:
    1. Arm CPU Architecture Protocol Dxe
    2. NOR Flash Dxe
    3. Runtime Variable Dxe

NvVarStoreFormattedLib was recently introduced to resolve the dependency
order. This patch propagates the necessary changes for ARM VExpress
CTA15+A7 platform and also fixes the "Firmware Volume for Variable Store
is corrupted" error seen when the Flash is erased (or not formatted).

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>
Signed-off-by: Alexei Fedorov <alexei.fedorov@arm.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>