Silicon/SynQuacer/PlatformDxe: depex on gEfiVariableArchProtocolGuid
SynQuacer's PlatformDxe uses HII style dynamic PCDs, and so it
implicitly depends on the PI variable protocol. This dependency is
not made explicit due to the fact that it is platform dependent
whether a certain variable is backed by such a dynamic PCD.

So add gEfiVariableArchProtocolGuid to PlatformDxe's DEPEX to ensure
that it does not attempt to refer to HII style dynamic PCDs before
the driver that produces them has been dispatched.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>