Silicon/AMD/Styx: remove MpBootDxe driver
MpBootDxe implements support for the ACPI Parking Protocol and the
DT based spintable protocol to bring up secondaries, neither of which
are recommended on systems implementing EL3, are not enabled in the
default build configuration of the platforms that include it and is
therefore essentially dead code. (Note that this driver DEPEXes on
a protocol that never gets installed when building this platform with
ARM-TF and PSCI support, which we now do unconditionally)

So let's remove it from the platform descriptions that refer to it,
and remove the code altogether as well.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>