Platform/DeveloperBox: switch to protocol based DevicePathLib
We already include DevicePathDxe, which exposes a protocol that implements
DevicePathLib in a way that can be shared across modules, avoiding the
need for multiple copies of UefiDevicePathLib, which is rather heavy
weight. However, we are not actually using it, and all DevicePathLib
users carry a private copy of UefiDevicePathLib.

So wire up all modules except DxeCore and DevicePathDxe itself, which
cannot use the protocol for obvious reasons. Also, avoid creating a
cyclic dependency between PcdDxe and DevicePathDxe by using PcdLibNull
in the latter.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>