Silicon/Socionext/SynQuacerPciHostBridgeLib: DEPEX on variable protocol
As discussed on-list [0], dynamic HII PCDs are backed by EFI variables,
but having PcdDxe DEPEXing on gEfiVariableArchProtocolGuid in general
is too restrictive, and so it is up to the platforms themselves to
DEPEX on gEfiVariableArchProtocolGuid in modules that use PCDs that
the platform may declare as dynamic HII. This also applies to the PCD
gSynQuacerTokenSpaceGuid.PcdPlatformSettings on the SynQuacer platform,
which is used by its PciHostBridgeLib implementation. So add the DEPEX
to make this dependency explicit.

[0] https://lists.01.org/pipermail/edk2-devel/2018-April/023700.html

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>