Silicon/Socionext/SynQuacer/Stage2Tables: add north SMMU level 3 table
Extend the static stage 2 page tables with a set of level 3 tables that
describe the ECAM space in a manner that allows the north SMMU to be used
to make the ECAM space appear sane to the CPUs.

It is up to the Secure firmware to manipulate the north SMMU page tables
so that the level 2 block entries corresponding with busses #0 .. #1 in
the respective config spaces of PCI0 and PCI1 are replaced with table
entries pointing to the level 3 tables added by this patch.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>