Hisilicon/Pci: move EnlargeAtuConfig0() to PciPlatformLib
we move all platform specific code to platform specific module, not
in PciHostBridge driver.
This is to prepare for switching to generic PciHostBridge driver, so

This patch is to move EnlargeAtuConfig0() into PcieInitDxe, in
PlatformNotify() of EFI_PCI_PLATFORM_PROTOCOL and implement api
PhaseNotifyPlatform for PciPlatform driver.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Heyi Guo <heyi.guo@linaro.org>
Signed-off-by: Ming Huang <ming.huang@linaro.org>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>