Marvell/Library: ComPhyLib: Configure SATA, SGMII and SFI in ARM-TF
Because it is now available to configure SerDes lanes via SiP
services in the EL3, there is no need to maintain same code
within UEFI.

This patch reworks serdes for: SATA, SGMII, HS-SGMII and SFI interfaces.
The next interfaces will be addressed in upcoming commits.
Parameters' description used in .dsc files and the
ComPhyLib were aligned to the ones used in EL3. For this purpose
and additional header was created.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>