Platform/ARM/Juno: increase max variable size to 8KB
Commit dc37ca75 ("Edk2Platforms: Replace MdeModulePkg PXE/iSCSI/TCP with
NetworkPkg drivers") switched to using iSCSI driver from the NetworkPkg
package. This driver requires the platform to support a maximum variable
size of atleast 4KB.

So increase the maximum supported variable size to 8KB on the Juno
platform. Without this, the iSCSI driver fails to load.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Reviewed-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Tested-by: Thomas Abraham <thomas.abraham@arm.com>