Silicon/Styx/AcpiPlatformDxe: simplify MADT generation logic
The MADT generation code is rather unpolished and overcomplicated,
and rewrites most of the table with information that is known
statically at build time, with the exception of which CPUs are
actually online.

So remove all the runtime generation code, and replace it with a
simple loop that sets the 'enabled' flag on each CPU that is
reported as present by the secure firmware.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>