Silicon/Styx: drop ARM_CPU_AARCH64 define
The define ARM_CPU_AARCH64 is only tested once, in the SMBIOS driver,
to decide whether to emit 'v8' or 'v7' as processor architecture.
However, this platform has no 32-bit addressable DRAM, and so it
cannot be built or executed in 32-bit mode anyway, making the test
rather pointless.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>