Silicon/Styx: switch to device path protocol driver
Replace the default resolution of the DevicePathLib library with
the version that invokes the DevicePathDxe protocol driver, which
was created to avoid having to carry a copy of the entire library
in each module.

Note that this driver itself incorporates on the NULL PcdLib instance,
to avoid creating a circular dependency on the PCD DXE driver. The
driver itself nor the DXE core can depend on the protocol, so they
keep using the library directly.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>