Platform/AMD/OverdriveBoard: build device tree from source
Instead of incorporating the prebuilt device tree binary as a
FREEFORM FFS file into the build, define a module that contains
the source so that the device tree compiler is invoked at build
time.

The original .dts file is moved into the new module, and cleaned
up a little so that explicit phandle properties are dropped, and
unused clocks and other redudant pieces are removed as well. The
existing prebuilt binary is deleted.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>