Silicon/AMD/Styx/Iort: drop conditionally included XGBE nodes
The I/O Remapping Table (IORT) does not require that each device of
which it describes the I/O remapping actually exists in the namespace.

So let's simplify the code by always including the XGBE nodes, and
just let them be unused if the platform does not include XGBE controllers.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>