Platform/FVP-AArch64: switch to the SBSA watchdog
On the FVP Foundation model, the SP805 watchdog appears to be 'wired'
incorrectly, resulting in a watchdog counter that decrements at the
APB clock rate of 24 MHz instead of the usual 32 kHz. Since the
timer start value is only 32-bits wide, this makes the watchdog
unusable in UEFI, since the default timeout set by the DXE core is
5 minutes, which is not representable in 32-bit at this clock rate.

So switch to the SBSA watchdog instead, which is wired up to the
generic timer, and ticks at the correct rate.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>